requirements.txt
colormath==3.0.0
networkx==2.8.7
numpy~=1.16.0
opencv-python==4.6.0.66
Pillow==9.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
idna==3.2
requests==2.26.0
selenium==3.141.0
urllib3==1.26.6
webdriver-manager==3.4.2
piexif~=1.1.3
imutils~=0.5.4
tinydb~=4.7.0
pypdf2~=2.11.2
nltk~=3.7
eyed3~=0.9.7
gpsphoto~=2.2.3
geopy~=2.3.0
customtkinter~=5.0.2
tkcalendar~=1.6.1
python-dateutil~=2.8.2
screeninfo~=0.8.1